Start hidden service in a new thread in the background & start ui. (#2)
Starting point with ui & hidden service together.

+++ priva/run.py
@@ -1,14 +1,12 @@
+import os
 from http.server import HTTPServer, BaseHTTPRequestHandler
+import shutil
 import json
 from stem.control import Controller
 from flask import Flask
-from stem.control import Controller
-from flask import Flask
 from priva_modules import ui
-
-if __name__ == "__main__":
-  import os
-import shutil
+from threading import Thread
+from colorama import Fore, Style
 
 app = Flask(__name__)
 
@@ -20,40 +18,51 @@ def index():
 def find_successor():
   return "<h1>Successor found!</h1>"
 
-print(' * Connecting to tor')
+onion_addr = None
+def start_server():
+  with Controller.from_port() as controller:
+    controller.authenticate()
 
-with Controller.from_port() as controller:
-  controller.authenticate()
+    print(' * Connecting to tor')
+    # All hidden services have a directory on disk. Lets put ours in tor's data
+    # directory.
 
-  # All hidden services have a directory on disk. Lets put ours in tor's data
-  # directory.
+    hidden_service_dir = os.path.join(controller.get_conf('DataDirectory', '/tmp'), 'priva')
 
-  hidden_service_dir = os.path.join(controller.get_conf('DataDirectory', '/tmp'), 'priva')
+    if os.path.isdir(hidden_service_dir):
+      controller.remove_hidden_service(hidden_service_dir)
+      shutil.rmtree(hidden_service_dir)
 
-  # Create a hidden service where visitors of port 80 get redirected to local
-  # port 5000 (this is where Flask runs by default).
+    # Create a hidden service where visitors of port 80 get redirected to local
+    # port 5000 (this is where Flask runs by default).
 
-  print(" * Creating our hidden service in %s" % hidden_service_dir)
-  result = controller.create_hidden_service(hidden_service_dir, 80, target_port = 5000)
-
-  try:
-    app.run()
+    print(" * Creating our hidden service in %s" % hidden_service_dir)
+    try:
+      result = controller.create_hidden_service(hidden_service_dir, 80, target_port = 5000)
+    except:
+      print(f"{Fore.RED}* Error: cannot start tor hidden service!{Style.RESET_ALL}")
     # The hostname is only available when we can read the hidden service
     # directory. This requires us to be running with the same user as tor.
 
     if result.hostname:
+      onion_addr = result.hostname
       print(" * Priva is available at %s, press ctrl+c to quit" % result.hostname)
-      UI.init_ui(result.hostname)
     else:
       print(" * Unable to determine our service's hostname, probably due to being unable to read the hidden service directory")
-    
-  finally:
-    # Shut down the hidden service and clean it off disk. Note that you *don't*
-    # want to delete the hidden service directory if you'd like to have this
-    # same *.onion address in the future.
 
-    print(" * Shutting down our hidden service")
-    controller.remove_hidden_service(hidden_service_dir)
-    shutil.rmtree(hidden_service_dir)
-      
-  app.run()
+    try:
+      app.run()
+    finally:
+      # Shut down the hidden service and clean it off disk. Note that you *don't*
+      # want to delete the hidden service directory if you'd like to have this
+      # same *.onion address in the future.
+
+      print(" * Shutting down our hidden service")
+      controller.remove_hidden_service(hidden_service_dir)
+      shutil.rmtree(hidden_service_dir)
+
+    app.run()
+
+# run the server in the background
+Thread(target=start_server)
+ui.UI.init_ui(onion_addr)
